- adding copy and clean for a theme/plugin build process
With copy we can...copy files. Yup totally random I know. This is going
to let us build a 'build' directory so we can ship a plugin/theme.

The clean module will clean up after our build so that we only ship the
things we need to. So remove the tests, scss,....whatever you want.